pyproject.toml
[tool.poetry]
name = "fastapi-rbac"
version = "0.1.0"
description = ""
authors = ["mnaimfaizy <mnaim.faizy@gmail.com>"]
readme = "README.md"
package-mode=false

[tool.black]
line-length = 88
target-version = [ "py310", "py311" ]
exclude = "((.eggs | .git | .pytest_cache | build | dist))"

[tool.ruff]
line-length = 88
exclude = [".git", "__pycache__", ".mypy_cache", ".pytest_cache"]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    # "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "B904",
    "B006",
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = ["alembic", "__pycache__"]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
fastapi = {extras = ["all"], version = "^0.110.2"}
python-multipart = "^0.0.9"
sqlalchemy = "^2.0.29"
alembic = "^1.13.1"
pandas = "^2.2.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pika = "^1.3.2"
python-crontab = "^3.0.0"
tenacity = "^8.2.3"
python-json-logger = "^2.0.7"
pydantic-settings = "^2.2.1"
sentry-sdk = "^1.45.0"
sqlalchemy-utils = "^0.41.2"
pydantic = {extras = ["email"], version = "^2.7.0"}
sqlmodel = "^0.0.16"
asyncpg = "^0.29.0"
cryptography = "^42.0.5"
asyncer = "^0.0.5"
httpx = "^0.27.0"
fastapi-async-sqlalchemy = "^0.6.1"
fastapi-pagination = {extras = ["sqlalchemy"], version = "^0.12.23"}
redis = "^5.0.3"
fastapi-cache2 = "^0.2.1"
fastapi-limiter = "^0.1.6"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
requests = "^2.31.0"
setuptools = "^69.5.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
flake8 = "^7.0.0"
black = {version = "^24.4.0", allow-prereleases = true}
autoflake = "^2.3.1"
isort = "^5.13.2"
sqlalchemy-stubs = "^0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
